% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Calc_Logistic_LAA}
\alias{Calc_Logistic_LAA}
\title{Calc_Logistic_LAA}
\usage{
Calc_Logistic_LAA(Linf, k, L0, Age)
}
\arguments{
\item{Linf}{A single value of asymptotic length for the logistic model}

\item{k}{A single value of the growth completion parameter for the logistic model}

\item{L0}{A single value of length-at-birth for the logistic model}

\item{Age}{A single value or vector of ages to convert to length based on the logistic model}
}
\value{
A vector of length-at-ages
}
\description{
Calc_Logistic_LAA
}
\examples{
# Calculate length-at-age for a logistic growth curve with given parameters
# and age range
Age_Range <- 0:16
Calc_Logistic_LAA(Linf = 1560, k = 0.36, L0 = 73, Age = Age_Range)
}
