% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Calc_Schnute_LAA}
\alias{Calc_Schnute_LAA}
\title{Calc_Schnute_LAA}
\usage{
Calc_Schnute_LAA(L2, a, L1, b, tau1, tau2, Age)
}
\arguments{
\item{L2}{A single value of large length for the Schnute model}

\item{a}{A single value of the growth completion parameter for the Schnute model}

\item{L1}{A single value of small length for the Schnute model}

\item{b}{A single value of the growth completion parameter for the Schnute model}

\item{tau1}{The age for the small length L2}

\item{tau2}{The age for the large length L1}

\item{Age}{A single value or vector of ages to convert to length based on the Schnute model}
}
\value{
A vector of length-at-ages
}
\description{
Calc_Schnute_LAA
}
