% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Calculate_DIC}
\alias{Calculate_DIC}
\title{Calculate_DIC}
\usage{
Calculate_DIC(data, Linf = NULL, Linf.se = NULL, L0 = NULL,
  L0.se = NULL, k.max = NULL, sigma.max = NULL, iter = 10000,
  BurnIn = 1000, n.chains = 4, thin = iter/10)
}
\arguments{
\item{data}{A data.frame that contains columns named 'Age' and "Length'. The function can
detect columns with similar names. If age and length columns cannot be determined then
an error will occur. The dataset can have additional columns which will be ignored by
the function}

\item{Linf}{The prior for asymptotic length. MUst be in the same unit (i.e. cm or mm) as the data.
This should be based off of maximum size for the species.}

\item{Linf.se}{The prior for normally distributed standard error around
asymptotic length. MUst be in the same unit (i.e. cm or mm) as the data. Cannot be zero.}

\item{L0}{The prior for length-at-birth. MUst be in the same unit (i.e. cm or mm) as the data.
This should be based off of minimum size for the species.}

\item{L0.se}{The prior for normally distributed standard error around
length-at-birth. MUst be in the same unit (i.e. cm or mm) as the data. Cannot be zero.}

\item{k.max}{The maximum value to consider for the growth completion parameter 'k'. In
the Gompertz and Logistic models, this parameter is often notated as 'g' instead of 'k'.}

\item{sigma.max}{The maximum value to consider for sigma. This is the variance around the
length-at-age residuals.}

\item{iter}{How many MCMC iterations should be run? Default is 10000 but fewer can be useful to
avoid longer run times when testing code or data}

\item{BurnIn}{The number of iterations at the begining of each chain to discard ('Burn in') to
avoid biased values from starting values that do not resemble the target distribution.
Default is 1000.}

\item{n.chains}{Number of MCMC chains to be run. Default is 4.}

\item{thin}{The thinning of the MCMC simulations. Deafault is 1 which means no thinning occurs.
Thinning is generally only necessary for complicated models as it increases run time.}
}
\value{
A data.frame with the results of the DIC analysis for each model.
}
\description{
Deviance Information Criteria (DIC) are calculated for three growth models
    (von Bertalanffy, Gompertz and Logistic) using the same prior parameters for each. These
    are used for model selection.
}
