% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Estimate_MCMC_BC_Growth}
\alias{Estimate_MCMC_BC_Growth}
\title{Estimate_MCMC_BC_Growth}
\usage{
Estimate_MCMC_BC_Growth(
  data,
  ID_col = NULL,
  Model = NULL,
  Linf = NULL,
  Linf.se = NULL,
  L0 = NULL,
  L0.se = NULL,
  k.max = NULL,
  sigma.max = NULL,
  iter = 10000,
  BurnIn = iter/2,
  n_cores = 1,
  controls = NULL,
  n.chains = 4,
  thin = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame that contains columns named 'Age' and "Length' as well as a column that specifies a ID or tag
that each observation belongs to. The function can detect columns with similar names for length and age.
If age and length columns cannot be determined then an error will occur. The ID column name is specified
through the `ID_col` variable. The dataset can have additional columns which will be ignored by
the function}

\item{ID_col}{The column name of the ID to be used for hierarchical structure (e.g. 'Tag', 'ID
'Sex', 'region', etc.)}

\item{Model}{Which growth model should be run? Must be one of "VB", "Gom" or "Log" for von
Bertalanffy, Gompertz or Logistic models, respectively}

\item{Linf}{The prior for asymptotic length. Must be in the same unit (i.e. cm or mm) as the data.
This should be based off of maximum size for the species.}

\item{Linf.se}{The prior for normally distributed standard error around
asymptotic length. Must be in the same unit (i.e. cm or mm) as the data. Cannot be zero.}

\item{L0}{The prior for length-at-birth. Must be in the same unit (i.e. cm or mm) as the data.
This should be based off of minimum size for the species.}

\item{L0.se}{The prior for normally distributed standard error around
length-at-birth. Must be in the same unit (i.e. cm or mm) as the data. Cannot be zero.}

\item{k.max}{The maximum value to consider for the growth completion parameter 'k'. In
the Gompertz and Logistic models, this parameter is often notated as 'g' instead of 'k'.}

\item{sigma.max}{The maximum value to consider for sigma. This is the variance around the
length-at-age residuals.}

\item{iter}{How many MCMC iterations should be run? Default is 10000 but fewer can be useful to
avoid longer run times when testing code or data}

\item{BurnIn}{The number of iterations at the beginning of each chain to discard ('Burn in') to
avoid biased values from starting values that do not resemble the target distribution.
Default is iter/2.}

\item{n_cores}{The number of cores to be used for parallel processing. It should be 1 core less than the
maximum number available.}

\item{controls}{A named list of parameters to control the rstan models behaviour.}

\item{n.chains}{Number of MCMC chains to be run. Default is 4.}

\item{thin}{The thinning of the MCMC simulations. Default is 1 which means no thinning occurs.
Thinning is generally only necessary for complicated models as it increases run time.}

\item{verbose}{TRUE or FALSE: flag indicating whether to print intermediate output from Stan on the console,
which might be helpful for model debugging.}
}
\value{
An object of class 'stanfit' from the rstan package.
}
\description{
A wrapper function that creates a Stan MCMC model using the rstan package.
    The data and priors provided are combined into an rstan model that estimates a hierarchical length-at-age
    model with a normal distribution. This allows simultaneous growth curves to be fit for Individuals based on
    tag numbers as well as at the population level. Three different growth models can
    be used: a von Bertalanffy model, Gompertz model or a logistic model. The prior on Linf and L0 are normally distributed
    and determined through the user providing a mean and se for each parameter. The growth completion
    parameter for any model (k) has a uniform prior which only requires an upper bound with the
    lower bound set at zero. Sigma is the residual variance of the data around the model and
    is set up in the same manner as 'k'. The growth estimates in the model are truncated to
    remain above zero so negative growth cannot occur.
}
\examples{
\donttest{
# load example data
data("example_BC_data")

# Use the function to estimate the rstan model
fit <- Estimate_MCMC_BC_Growth(data = example_BC_data,
                                  ID_col = "Tag",
                                  Model = "VB" ,
                                  iter = 500,
                                  Linf = 2400,
                                  Linf.se = 50,
                                  L0 = 660,
                                  sigma.max = 200,
                                  L0.se = 67,
                                  k.max = 1,
                                  verbose = TRUE,
                                  n_cores = 1)
}
}
