% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Estimate_MCMC_Group_Growth}
\alias{Estimate_MCMC_Group_Growth}
\title{Estimate_MCMC_Group_Growth}
\usage{
Estimate_MCMC_Group_Growth(
  data,
  group_col = NULL,
  Model = NULL,
  Linf = NULL,
  Linf.se = NULL,
  L0 = NULL,
  L0.se = NULL,
  k.max = NULL,
  sigma.max = NULL,
  iter = 10000,
  BurnIn = iter/2,
  n_cores = 1,
  controls = NULL,
  n.chains = 4,
  thin = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{Text}

\item{group_col}{Text}

\item{Model}{Text}

\item{Linf}{Text}

\item{Linf.se}{Text}

\item{L0}{Text}

\item{L0.se}{Text}

\item{k.max}{Text}

\item{sigma.max}{Text}

\item{iter}{Text}

\item{BurnIn}{Text}

\item{n_cores}{Text}

\item{controls}{Text}

\item{n.chains}{Text}

\item{thin}{Text}

\item{verbose}{Text}
}
\value{
Text
}
\description{
A wrapper function that creates a Stan MCMC model using the rstan package.
}
