% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGrowth.R
\docType{data}
\name{example_BC_data}
\alias{example_BC_data}
\title{Example back-calculated length-at-age data}
\format{
A data frame with 294 rows and 4 variables
}
\usage{
data(example_BC_data)
}
\description{
A dataset used as examples in the vignettes and for users to test code with
}
\details{
\itemize{
  \item Age. Number of growth bands determined from vertebral analysis
  \item Length. Total Length in mm determined via back-calculation
  \item Sex. F for female and M for male
  \item Tag. ID of each individual for back-calculated length-at-age
}
}
\keyword{datasets}
