% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGrowth.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example length-at-age data}
\format{
A data frame with 509 rows and 2 variables
}
\usage{
data(example_data)
}
\description{
A dataset used as examples in the vignettes and for users to test code with
}
\details{
\itemize{
  \item Age. Number of growth bands determined from otolith ageing
  \item Length. Total Length in mm determined
}
}
\keyword{datasets}
