% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGrowth.R
\docType{data}
\name{Looic_example_results}
\alias{Looic_example_results}
\title{Example MCMC model comparison outputs outputs}
\format{
An dataframe with results of LooIC
}
\usage{
data(Looic_example_results)
}
\description{
The results of an MCMC model comparison used as examples in the vignettes and for users to test code with
}
\keyword{datasets}
