% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGrowth.R
\docType{data}
\name{MCMC_example_results}
\alias{MCMC_example_results}
\title{Example MCMC model outputs}
\format{
An rstan model with the class stan.fit
}
\usage{
data(MCMC_example_results)
}
\description{
The results of an MCMC model used as examples in the vignettes and for users to test code with
}
\keyword{datasets}
