% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGrowth.R
\docType{package}
\name{BayesGrowth}
\alias{BayesGrowth}
\alias{BayesGrowth-package}
\title{BayesGrowth}
\description{
A package to estimate length-at-age using MCMC using rjags models
}
\author{
Jonathan Smart
}
