% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Get_MCMC_parameters}
\alias{Get_MCMC_parameters}
\title{Get_MCMC_parameters}
\usage{
Get_MCMC_parameters(data)
}
\arguments{
\item{data}{n output from the Estimate_MCMC_Growth funtion or a similarly structure rjags output.
Must be of class "mcmc.list"}
}
\value{
A tibble with 5 columns that include each parameter and the results of their posterior distributions.
    These include the mean, Standard deviation, lower 95th percentile and upper 95th percentile
}
\description{
Get parameter summary statistics from the outputs of a Estimate_MCMC_Growth object
}
