% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Get_MCMC_parameters}
\alias{Get_MCMC_parameters}
\title{Get_MCMC_parameters}
\usage{
Get_MCMC_parameters(obj)
}
\arguments{
\item{obj}{An output from the Estimate_MCMC_Growth function}
}
\value{
A data.frame with the posterior distributions for each parameter.
    These include the mean, Standard error of the mean, Standard deviationof the mean, median,
    95th percentiles, effective sample sizes and Rhat.
}
\description{
Get parameter summary statistics from the outputs of a Estimate_MCMC_Growth object. It is simplified set of
    results than is returned from summary(obj).
}
